+++ ADCSAlgorithms/messaging/static_messaging.h
@@ -15,7 +15,7 @@
 
    void InitializeStorage(uint32_t StorageBytes);
    int32_t CreateNewMessage(char* MessageName, uint32_t MaxSize);
-   int32_t WriteMessage(uint32_t MessageID, uint32_t ClockTimeNanos, uint32_t MsgSize,
+   int32_t WriteMessage(uint32_t MessageID, uint64_t ClockTimeNanos, uint32_t MsgSize,
       void *MsgPayload);
    int32_t ReadMessage(uint32_t MessageID, uint64_t *WriteTime, uint32_t *WriteSize,
       uint32_t MaxBytes, void *MsgPayload);
